+++ src/ui/ScreenList.cpp
@@ -22,21 +22,21 @@ vui::IGameScreen* vui::ScreenList::movePrevious() {
 void vui::ScreenList::setScreen(i32 s) {
     _current = s;
 }
-vui::ScreenList* vui::ScreenList::addScreens(IGameScreen** s, i32 c /*= 0*/) {
+vui::ScreenList* vui::ScreenList::addScreens(IGameScreen** s, size_t c /*= 0*/) {
     // Check For Arguments
     if (s == 0 || c < 1) return this;
 
     // Copy Over The Screens
-    int cs = _screens.size(), ds = cs + c;
+    size_t cs = _screens.size(), ds = cs + c;
     _screens.resize(ds);
-    for (int i = 0; i < c; i++) {
+    for (size_t i = 0; i < c; i++) {
         _screens[cs + i] = s[i];
     }
 
     // Build The Added Screens
-    for (int i = cs; i < ds; i++) {
+    for (size_t i = cs; i < ds; i++) {
         if (_screens[i] == 0) continue;
-        _screens[i]->setParentGame(_game, i);
+        _screens[i]->setParentGame(_game, (i32)i);
         _screens[i]->build();
     }
 
@@ -45,16 +45,16 @@ vui::ScreenList* vui::ScreenList::addScreens(IGameScreen** s, i32 c /*= 0*/) {
 vui::ScreenList* vui::ScreenList::addScreen(IGameScreen* s) {
     _screens.push_back(s);
     if (s != 0) {
-        s->setParentGame(_game, _screens.size() - 1);
+        s->setParentGame(_game, (i32)_screens.size() - 1);
         s->build();
     }
     return this;
 }
 
 void vui::ScreenList::destroy(GameTime gameTime) {
-    int s = _screens.size();
+    i32 s = (i32)_screens.size();
     if (s < 1) return;
-    for (int i = 0; i < s; i++) {
+    for (i32 i = 0; i < s; i++) {
         _screens[i]->destroy(gameTime);
     }
     _screens.resize(0);
